requirements.txt
Django==5.0.7
djangorestframework==3.15.2
Django==5.0.7
djangorestframework==3.15.2
whitenoise==6.7.0
gunicorn==21.2.0